Reorganize test database setup files
Refactor the organization of the test database setup scripts and related files from the `test_database_setup` directory to the `database` directory. This restructuring simplifies file navigation, aligns with the project's structure and it is easier for team members contributing to the code in the future. The files moved include the Docker setup, shell scripts and README files.