Add main.py for google searcher application
This commit introduces a new python file - main.py within the google_searcher module. This file is responsible for initiating the GoogleSearchQueueManager, GoogleSearcher, and DBManager objects, thereby managing the Google search API calls based on the quota. It is designed to stop the search operation once the allocated quota is exhausted. This file also integrates with the DigitalOcean database for result collection and utilizes environment variables for secure and smooth operation.