Remove automatic PR commenting script
This commit deletes the script responsible for automatically commenting on Pull Requests with the results of 'mypy' and 'pydocstyle' checks. Additionally, references to this script have been removed from the 'python_checks.yml' GitHub workflow file. As a result, the GitHub actions workflow now only performs the checks without posting comments, simplifying the review process and reducing unnecessary noise in the PR discussion.