Optimize script for finding modified Python files
This commit simplifies the method of detecting modified Python files in the GitHub Actions workflow file 'python_checks.yml'. Instead of writing the output of 'git diff' command to the environment variable 'MODIFIED_FILES' using multiple lines, it now does so in a single line, thereby improving readability and maintaining cleanliness in the code.