Refactor Github action workflows to handle errors
The Python checks in the Github action workflow for mypy and pydocstyle have been updated to capture their respective exit codes. An additional step has been included to observe and react to those exit codes, terminating the workflow with a failure status if any errors are detected in either of the checks.